Normalize the capitalization of common names
I've only touched the "name" field so far—the "description" field still needs to be updated. (I'll get to that once I do a few other normalizations.)

The capitalization scheme we settled on was "entirely lowercase except for proper nouns."

When in doubt about whether a word was proper, I looked at how people on the internet capitalized it. For example, it looks like "spotted Joe Pye weed" is proper (named after a real person, a Native American doctor from the 1700s, see https://en.wikipedia.org/wiki/Eutrochium) while "alexanders" is not (probably originally named after Alexander the Great, but used for long enough that it's a common noun now, see https://en.wiktionary.org/wiki/alexanders). There seems to be a bit of variation with both of these, however.