In the schemas, enforce a pattern for scientific names
The JSON data already conforms to this schema rule.

We can't use a pattern to describe the common names, not even to make sure they're lowercase, because there's no way to tell what's a proper noun and what isn't.

However, scientific names have a regular enough structure that we can require a certain formatting from them. (In particular, they must consist of two words, space-separated, with the first word capitalized and the second word entirely lowercase.)

Getting a bit deep in the weeds (pun intended), it looks like scientific names of plants are allowed to contain 1) plain ASCII latin letters, 2) hyphens, and 3) vowels with dieareses.

(The New England aster, Symphyotrichum novae-angliae, is an example of rule 2; the lake quillwort, Isoëtes lacustris, is an example of rule 3.)

See https://www.iapt-taxon.org/nomen/pages/main/art_60.html for rules on the precise structure of scientific names.