Use less latin in bees' pollen_collection field
In particular, I've decided to replace the words "femur", "tibia", and "basitarsus"/"basitarsis".

I replaced "Femur" and "tibia" with "leg"; I replaced the phrase "tibia and basitarsus" with "lower leg". I know that isn't quite as specific, but it's probably more informative to students who may or may not know what a tibia is. (I had to look up what a tibia was just now! 🙃)

I decided to keep "abdomen", since that's a familiar English word, and also "thorax", because I don't know of a good equivalent. (Possibly "torso"?)

This commit also makes a few sentence structure changes. (Most of these are changes from the passive voice to the active voice; I also reworded one or two sentences for better flow.)