Remove a note about mason bee's pollen provisions
"Provisions" is a sufficiently obscure word that students probably wouldn't understand what it means. (In fact, I didn't understand what it meant until Paul googled it.)

I still don't fully understand what this sentence means, so I don't think I could rewrite it to be more clear. It doesn't seem to be a critical piece of information, though, so I think we can just delete it. :)